+++ packages/backend-core/src/events/processors/async/DocumentUpdateProcessor.ts
@@ -1,12 +1,11 @@
 import { EventProcessor } from "../types"
 import { Event, Identity, DocUpdateEvent } from "@budibase/types"
-import { DocumentType, SEPARATOR } from "../../../constants"
 import { doInTenant } from "../../../context"
 import { getDocumentId } from "../../documentId"
 import { shutdown } from "../../asyncEvents"
 
 export type Processor = (update: DocUpdateEvent) => Promise<void>
-export type ProcessorMap = { types: DocumentType[]; processor: Processor }[]
+export type ProcessorMap = { events: Event[]; processor: Processor }[]
 
 export default class DocumentUpdateProcessor implements EventProcessor {
   processors: ProcessorMap = []
@@ -26,8 +25,8 @@ export default class DocumentUpdateProcessor implements EventProcessor {
     if (!tenantId || !docId) {
       return
     }
-    for (let { types, processor } of this.processors) {
-      if (types.find(type => docId.startsWith(`${type}${SEPARATOR}`))) {
+    for (let { events, processor } of this.processors) {
+      if (events.includes(event)) {
         await doInTenant(tenantId, async () => {
           await processor({
             id: docId,
