+++ packages/backend-core/src/queue/constants.ts
@@ -2,5 +2,5 @@ export enum JobQueue {
   AUTOMATION = "automationQueue",
   APP_BACKUP = "appBackupQueue",
   AUDIT_LOG = "auditLogQueue",
-  DOC_UPDATE = "docUpdateQueue",
+  SYSTEM_EVENT_QUEUE = "systemEventQueue",
 }
