+++ packages/server/src/events/docUpdates/index.ts
@@ -1,0 +1,1 @@
+export * from "./processors"
